requirements.txt
google-generativeai
flask
python-dotenv
SpeechRecognition
whisper
pydub